chore: moving test helpers to the root of the monorepo
