+++ packages/ui-components/src/components/Bubble/__tests__/Bubble.test.tsx
@@ -3,7 +3,7 @@
 import { act, fireEvent, render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 
-import { expectToHaveClasses } from "../../../common/__tests__/helpers";
+import { expectToHaveClasses } from "../../../../../../configuration/tests-helpers";
 import { BUBBLE_CLASSNAME } from "../../../common/constants";
 import { Bubble } from "../..";
