+++ packages/ui-components/src/components/Button/__tests__/ButtonLink.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 
-import { expectToHaveClasses } from "../../../common/__tests__/helpers";
+import { expectToHaveClasses } from "../../../../../../configuration/tests-helpers";
 import { ButtonLink } from "../..";
 
 describe("ButtonLink (exceptions)", () => {
