+++ packages/ui-components/src/components/Card/__tests__/Card.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 
-import { expectToHaveClasses } from "../../../common/__tests__/helpers";
+import { expectToHaveClasses } from "../../../../../../configuration/tests-helpers";
 import { CARD_CLASSNAME } from "../../../common/constants";
 import { Card } from "../..";
