+++ packages/ui-components/src/components/Footer/__tests__/Footer.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 
-import { expectToHaveClasses } from "../../../common/__tests__/helpers";
+import { expectToHaveClasses } from "../../../../../../configuration/tests-helpers";
 import { FOOTER_CLASSNAME } from "../../../common/constants";
 import { Footer } from "../..";
