+++ packages/ui-components/src/components/Header/__tests__/Header.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 
-import { expectToHaveClasses } from "../../../common/__tests__/helpers";
+import { expectToHaveClasses } from "../../../../../../configuration/tests-helpers";
 import { HEADER_CLASSNAME } from "../../../common/constants";
 import { Header } from "../..";
