+++ packages/ui-components/src/components/Panel/__tests__/Panel.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 
-import { expectToHaveClasses } from "../../../common/__tests__/helpers";
+import { expectToHaveClasses } from "../../../../../../configuration/tests-helpers";
 import {
 	MESSAGEBOX_CLASSNAME,
 	PANEL_CLASSNAME,
