+++ packages/ui-components/src/components/Pill/__tests__/Pill.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 
-import { expectToHaveClasses } from "../../../common/__tests__/helpers";
+import { expectToHaveClasses } from "../../../../../../configuration/tests-helpers";
 import { PILL_CLASSNAME } from "../../../common/constants";
 import { Pill } from "../..";
