+++ packages/ui-components/src/components/Spinner/__tests__/Spinner.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 
-import { expectToHaveClasses } from "../../../common/__tests__/helpers";
+import { expectToHaveClasses } from "../../../../../../configuration/tests-helpers";
 import { SPINNER_CLASSNAME } from "../../../common/constants";
 import { Spinner } from "../..";
