+++ packages/ui-components/src/components/Table/__tests__/Table.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from "@testing-library/react";
 import {
 	expectToHaveClasses,
 	expectToHaveStyles,
-} from "../../../common/__tests__/helpers";
+} from "../../../../../../configuration/tests-helpers";
 import { Table, TableBody, TableCell, TableHead, TableRow } from "../..";
 
 describe("Table (exceptions)", () => {
